requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.26.106
botocore==1.29.106
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
idna==3.4
igraph==0.10.1
imageio==2.26.0
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
jupyterthemes==0.20.0
kiwisolver==1.4.4
lazy_loader==0.1
lesscpy==0.15.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mca==1.0.3
mistune==2.0.4
multidict==6.0.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
plotly==5.13.1
ply==3.11
prince==0.7.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.4
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.0
scipy==1.9.1
seaborn==0.12.1
selenium==4.9.1
Send2Trash==1.8.0
sfrancia==1.0.8
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
stack-data==0.6.2
statsmodels==0.13.5
stepwise-process==2.5
tenacity==8.2.2
terminado==0.17.1
texttable==1.6.4
threadpoolctl==3.1.0
tifffile==2023.2.28
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.5
webcolors==1.12
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.4
wsproto==1.2.0
yarl==1.8.2
zipp==3.11.0
